+++ fotoobo/cli/main.py
@@ -6,7 +6,10 @@ more than a few subcommands.
 
 Caution: Use docstrings with care as they are used to print help texts on any command.
 """
+# pylint: disable=anomalous-backslash-in-string
+
 import logging
+import os
 import sys
 from typing import Optional, Union
 
@@ -23,7 +26,11 @@ from fotoobo.helpers.config import config
 from fotoobo.helpers.log import Log
 from fotoobo.helpers.output import print_logo
 
-app = typer.Typer(context_settings={"help_option_names": ["-h", "--help"]}, no_args_is_help=True)
+app = typer.Typer(
+    context_settings={"help_option_names": ["-h", "--help"]},
+    no_args_is_help=True,
+    rich_markup_mode="rich",
+)
 log = logging.getLogger("fotoobo")
 
 
@@ -41,28 +48,34 @@ def callback(  # pylint: disable=too-many-arguments
         None,
         "--config",
         "-c",
-        help="The fotoobo configuration file",
+        help="Set the fotoobo configuration file. \[default: fotoobo.yaml]",
         show_default=False,
         metavar="[file]",
     ),
-    log_switch: bool = typer.Option(None, "--log", "-l", help="Enable logging"),
+    log_switch: bool = typer.Option(
+        None, "--log", "-l", help="Enable logging. \[default: disabled]"
+    ),
     log_level: str = typer.Option(
         None,
         "--loglevel",
-        help="The logging level. Choose from CRITICAL, ERROR, WARNING, INFO, DEBUG",
+        help="Set the log level. Choose from CRITICAL, ERROR, WARNING, INFO, DEBUG.",
         metavar="[level]",
+        show_default=False,
     ),
-    nologo: bool = typer.Option(None, "--nologo", help="Disable the logo"),
+    nologo: bool = typer.Option(None, "--nologo", help="Do not print the beautiful fotoobo logo."),
     version: Optional[bool] = typer.Option(  # pylint: disable=unused-argument
-        None, "--version", "-V", help="Get fotoobo version", callback=version_callback
+        None, "--version", "-V", help="Print the fotoobo version.", callback=version_callback
     ),
 ) -> None:
     """
     The Fortinet Toolbox (fotoobo) - make IT easy
 
-    This tool lets you interact with Fortinet assets like devices and their configuration.
+    ─────────────────────────────────────────────
 
-    For help see the README.md or theAttribute specific documentation in the docs folder.
+    This is fotoobo, the mighty [bold]Fo[/bold]rtinet [bold]too[/bold]l[bold]bo[/bold]x for
+    managing your Fortinet environment. It is meant to be extendable to your needs.
+
+    For detailed documentation see https://fotoobo.readthedocs.io.
     """
     config.load_configuration(config_file)
     config.no_logo = True if nologo else config.no_logo
@@ -96,24 +109,26 @@ def callback(  # pylint: disable=too-many-arguments
 @app.command(hidden=True)
 def greet(
     name: Optional[str] = typer.Argument(
-        "", help="The name of the person to greet", show_default=False, metavar="[name]"
+        None, help="The name of the person to greet.", show_default=False, metavar="[name]"
     ),
-    bye: bool = typer.Option(False, "--bye", "-b", help="Also write bye at the end"),
-    log_enabled: bool = typer.Option(False, "--log", "-l", help="Enable logging"),
+    bye: bool = typer.Option(False, "--bye", "-b", help='Also write "bye" at the end.'),
+    log_enabled: bool = typer.Option(False, "--log", "-l", help="Enable logging."),
 ) -> None:
     """
-    This is the hidden Greeting function.
+    This is the hidden greeting function.
     It allows you to greet someone with different colors in different languages.
     """
+    if not name:
+        name = os.getlogin().capitalize()
     utils.greet(str(name), bye, log_enabled)
 
 
 # fotoobo specific commands
-app.add_typer(convert.app, name="convert", help="Convert commands for fotoobo")
-app.add_typer(get.app, name="get", help="get commands for fotoobo")
+app.add_typer(convert.app, name="convert", help="Convert commands for fotoobo.")
+app.add_typer(get.app, name="get", help="Get information about fotoobo or your configuration.")
 
 # commands for the Fortinet products
-app.add_typer(ems.app, name="ems", help="Commands for FortiClient EMS")
-app.add_typer(faz.app, name="faz", help="Commands for FortiAnalyzer")
-app.add_typer(fgt.app, name="fgt", help="Commands for FortiGate")
-app.add_typer(fmg.app, name="fmg", help="Commands for FortiManager")
+app.add_typer(ems.app, name="ems", help="Commands for FortiClient EMS.")
+app.add_typer(faz.app, name="faz", help="Commands for FortiAnalyzer.")
+app.add_typer(fgt.app, name="fgt", help="Commands for FortiGate.")
+app.add_typer(fmg.app, name="fmg", help="Commands for FortiManager.")