+++ tests/cli/fgt/test_cli_fgt_config.py
@@ -21,7 +21,10 @@ def test_cli_app_fgt_config_help() -> None:
     assert set(commands) == {"check", "info"}
 
 
-def test_cli_app_fgt_config_check_no_args() -> None:
+def test_cli_app_fgt_config_no_args() -> None:
     """Test fgt config check with no arguments"""
-    result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fgt", "config", "check"])
-    assert result.exit_code == 2
+    result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fgt", "config"])
+    assert result.exit_code == 0
+    assert "Usage: callback fgt config [OPTIONS] COMMAND" in result.stdout
+    assert "--help" in result.stdout
+    assert "FortiGate config file commands." in result.stdout