+++ CHANGELOG.md
@@ -10,6 +10,12 @@ For examples and guidelines see [https://keepachangelog.com/](https://keepachang
 
 ### Changed
 
+- Renamed the following commands:
+   - `fotoobo fmg set` -> `fotoobo fmg post`
+   - `fotoobo fgt check` -> `fotoobo fgt monitor`
+- Changed argument ordering of the following commands (since host is usually optional):
+   - `fotoobo fmg assign`
+   - `fotoobo fmg get policy`
 - Invoking fotoobo (or any sub-command) without arguments displays the help instead of an error